Added safe way of parsing string to doubles for third party api payloads

+++ server/src/main/java/com/Adi/stock_tracker/dto/CompanyOverviewResponse.java
@@ -15,10 +15,10 @@ public record CompanyOverviewResponse(
         String country,
         String sector,
         String officialSite,
-        Long marketCap,
-        Long peRatio,
-        Long fiftyTwoWeekHigh,
-        Long fiftyTwoWeekLow
+        Double marketCap,
+        Double peRatio,
+        Double fiftyTwoWeekHigh,
+        Double fiftyTwoWeekLow
 
 ) {
 }
